requirements.txt
redis
boto3
pyjwt
jsonify
flask
requests
SQLAlchemy
